Fixed setBalance updates throughout. Added a way to display both balances inside Settings.
